Latest changes required to work on macOS

+++ setup.sh
@@ -37,13 +37,23 @@ export PYENV_ROOT="$HOME/.pyenv"
 
 if [ ! -d "$PYENV_ROOT" ]; then
 	git clone https://github.com/pyenv/pyenv.git "$PYENV_ROOT"
+	git clone https://github.com/pyenv/pyenv-update.git "$PYENV_ROOT/plugins/pyenv-update"
 	git clone https://github.com/pyenv/pyenv-virtualenv.git "$PYENV_ROOT/plugins/pyenv-virtualenv"
 	export PATH="$PYENV_ROOT/bin:$PATH"
-	DEFAULT_PYTHON_VERSION=$(pyenv install --list | grep -v - | grep -v b | grep -v mini | grep -v rc | tail -1 | awk '{ print $1 }')
-	pyenv install "$DEFAULT_PYTHON_VERSION"
-	pyenv global "$DEFAULT_PYTHON_VERSION"
+	if [ -f .python-version ]; then
+		pyenv install
+		pyenv global "$(cat .python-version)"
+	elif [ -f "$DOTFILES_DIR/.python-version" ]; then
+		pyenv install "$(cat "$DOTFILES_DIR/.python-version")"
+		pyenv global "$(cat "$DOTFILES_DIR/.python-version")"
+	else
+		DEFAULT_PYTHON_VERSION=$(pyenv install --list | grep -v - | grep -v a | grep -v b | grep -v mini | grep -v rc | tail -1 | awk '{ print $1 }')
+		pyenv install "$DEFAULT_PYTHON_VERSION"
+		pyenv global "$DEFAULT_PYTHON_VERSION"
+	fi
 	eval "$(pyenv init --path)"
 	eval "$(pyenv init -)"
+	# eval "$(pyenv virtualenv-init -)"
 	pip install --upgrade pip pip-tools
 	pip-sync "$DOTFILES_DIR/requirements.txt" "$DOTFILES_DIR/requirements-dev.txt"
 else
@@ -55,14 +65,25 @@ fi
 # Check for existing Python virtualenv called ansible-system
 CURRENT_PYTHON_VERSION=$(pyenv version | awk '{ print $1 }')
 set +e
-pyenv versions | grep ansible-system
-EXIT="$?"
-if [[ "$EXIT" == 1 ]]; then
-	# pyenv global system
-	pyenv virtualenv --system-site-packages system ansible-system
+if [[ $(uname) == "Linux" ]]; then
+	pyenv versions | grep ansible-system
+	EXIT="$?"
+	if [[ "$EXIT" == 1 ]]; then
+		# pyenv global system
+		pyenv virtualenv --system-site-packages system ansible-system
+	fi
+	pyenv global ansible-system
+else
+	pyenv versions | grep ansible
+	EXIT="$?"
+	if [[ "$EXIT" == 1 ]]; then
+		# pyenv global system
+		pyenv virtualenv "$(cat "$DOTFILES_DIR/.python-version")" ansible
+	fi
+	pyenv global ansible
 fi
+
 set -e
-pyenv global ansible-system
 pip3 install --upgrade pip
 pip3 install -r "$BUILD_DIR/requirements.txt" -r "$BUILD_DIR/requirements-dev.txt"
 cd "$BUILD_DIR" || exit
